Added RelativeError for MG Problems
Given that the multigroup problems have a "true" solution for scalar flux
when those problems are run the algorithm calculates the relative error and
stores that in the output file.

In the future I need to figure out how to make it do that based on whether or not it has a true_flux variable, not whether or not its a multigroup problem.