Fixed MG Problem
MG problems have not been working properly for awhile. Needed to transpose Sigs when calculating the source
in the Source Iteration. Reeds problem still works and I would suspect so does Garcia, but these are both 1G
problems. I am not sure if this fix is general or just fixing the arrangement of the HDPE data in particular.